Refactor API handlers: Separate types, transformers, and handlers into distinct files for improved maintainability and organization. Implement domain-specific handler files for channels, configuration, videos, and newsletters. Introduce a service container for dependency management. Remove legacy code and enhance testability by isolating concerns.
